requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.10.5
click==8.3.0
decouple==0.0.7
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.3
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.41.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
charset_normalizer==3.4.3
finnhub-python==2.4.25 
requests==2.32.5